+++ src/CoreEx/ConflictException.cs
@@ -15,11 +15,12 @@ namespace CoreEx
     public class ConflictException : Exception, IExtendedException
     {
         private const string _message = "A data conflict occurred.";
+        private static bool? _shouldExceptionBeLogged;
 
         /// <summary>
         /// Get or sets the <see cref="ShouldBeLogged"/> value.
         /// </summary>
-        public static bool ShouldExceptionBeLogged { get; set; }
+        public static bool ShouldExceptionBeLogged { get => _shouldExceptionBeLogged ?? Internal.ShouldExceptionBeLogged<ConflictException>(); set => _shouldExceptionBeLogged = value; }
 
         /// <summary>
         /// Initializes a new instance of the <see cref="ConflictException"/> class.