+++ src/CoreEx/DuplicateException.cs
@@ -14,11 +14,12 @@ namespace CoreEx
     public class DuplicateException : Exception, IExtendedException
     {
         private const string _message = "A duplicate error occurred.";
+        private static bool? _shouldExceptionBeLogged;
 
         /// <summary>
         /// Get or sets the <see cref="ShouldBeLogged"/> value.
         /// </summary>
-        public static bool ShouldExceptionBeLogged { get; set; }
+        public static bool ShouldExceptionBeLogged { get => _shouldExceptionBeLogged ?? Internal.ShouldExceptionBeLogged<DuplicateException>(); set => _shouldExceptionBeLogged = value; }
 
         /// <summary>
         /// Initializes a new instance of the <see cref="DuplicateException"/> class.