+++ src/CoreEx/NotFoundException.cs
@@ -14,11 +14,12 @@ namespace CoreEx
     public class NotFoundException : Exception, IExtendedException
     {
         private const string _message = "Requested data was not found.";
+        private static bool? _shouldExceptionBeLogged;
 
         /// <summary>
         /// Get or sets the <see cref="ShouldBeLogged"/> value.
         /// </summary>
-        public static bool ShouldExceptionBeLogged { get; set; }
+        public static bool ShouldExceptionBeLogged { get => _shouldExceptionBeLogged ?? Internal.ShouldExceptionBeLogged<NotFoundException>(); set => _shouldExceptionBeLogged = value; }
 
         /// <summary>
         /// Initializes a new instance of the <see cref="NotFoundException"/> class.